requirements.txt
google-adk>=0.1.0
requests>=2.0.0
python-dotenv>=1.1.0
